Move more code out of mysqlserver class
